Add collision detection & change rendering
Objects used to render cubes used to get a Camera object for
determinating view matrix, now the Camera class is canceled. At first I
wanted the rendering objects to have a Player object instead, but this
results in cyclic including for the header files. So I created the
RenderWorld class, but it will be canceled later.